Code review changes.
Also moved video call related code from topic.go to calls.go.
