+++ server/datamodel.go
@@ -296,8 +296,6 @@ type MsgClientDel struct {
 
 // MsgClientNote is a client-generated notification for topic subscribers {note}.
 type MsgClientNote struct {
-	Id string `json:"id,omitempty"`
-
 	Topic string `json:"topic"`
 	// what is being reported: "recv" - message received, "read" - message read, "kp" - typing notification
 	What string `json:"what"`
